Updated returns calculation to account for number of shares of the asset being sold.
Removed the parameter of a request user from the Transaction constructor. It is unnecessary to require a requested when first creating a transaction. Requested user set in the requestTransaction() method.

Fixed: setter method for requestUser was setting the resolveUser, not the requestUser

Changed: setReqUser to setRequestUser

Added: in Testing.java, added .requestTransaction() after each transaction instantiation to ensure the requestUser is set before getting it.

Changed: relocated the setting of the requestDate, setting of the transactionID, and the adding of a transaction to the transactionRequest list from the Transaction constructor and to requestTransaction().

Fixed: selling of assets correctly updates Bank balance; had to move the setting of finalPrice in Testing.java to before it the transaction is resolved.

Added: to Testing.java, verification that timestamps of Portfolio and Bank balances match with the corresponding associated transaction, if one exists (buying/selling only).
Signed-off-by: Ryan Tatton <rdt46@case.edu>